// order-service/src/shared/proto/order.ts
export interface RegisterOrderRequestAddress {
  address: string;
  neighborhood: string;
  number: string;
  complement: string;
  city: string;
  state: string;
  zipCode: string;
  latitude: number;
  longitude: number;
}

export interface RegisterOrderRequest {
  userId: string;
  distance: number;
  comment?: string | undefined;
  paymentMethod?: string | undefined;
  address: RegisterOrderRequestAddress | undefined;
}

export interface OrderResponse {
  message: string;
  order: Order | undefined;
}

export interface Address {
  id: string;
  orderId: string;
  address: string;
  neighborhood: string;
  number: string;
  complement: string;
  city: string;
  state: string;
  zipCode: string;
  latitude: number;
  longitude: number;
  createdAt: string;
  updatedAt: string;
}

export interface Order {
  id: string;
  userId: string;
  userName: string;
  userLastName?: string | undefined;
  userImage?: string | undefined;
  deliverymanId?: string | undefined;
  deliverymanName?: string | undefined;
  deliverymanLastName?: string | undefined;
  deliverymanImage?: string | undefined;
  comment?: string | undefined;
  value: number;
  discount?: number | undefined;
  tax?: number | undefined;
  total: number;
  status: string;
  paymentMethod: string;
  createdAt: string;
  updatedAt: string;
  address: Address | undefined;
}

export const ORDER_PACKAGE_NAME = "order";

function createBaseRegisterOrderRequestAddress(): RegisterOrderRequestAddress {
  return {
    address: "",
    neighborhood: "",
    number: "",
    complement: "",
    city: "",
    state: "",
    zipCode: "",
    latitude: 0,
    longitude: 0,
  };
}

export const RegisterOrderRequestAddress: MessageFns<RegisterOrderRequestAddress> = {
  encode(message: RegisterOrderRequestAddress, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.address !== "") {
      writer.uint32(10).string(message.address);
    }
    if (message.neighborhood !== "") {
      writer.uint32(18).string(message.neighborhood);
    }
    if (message.number !== "") {
      writer.uint32(26).string(message.number);
    }
    if (message.complement !== "") {
      writer.uint32(34).string(message.complement);
    }
    if (message.city !== "") {
      writer.uint32(42).string(message.city);
    }
    if (message.state !== "") {
      writer.uint32(50).string(message.state);
    }
    if (message.zipCode !== "") {
      writer.uint32(58).string(message.zipCode);
    }
    if (message.latitude !== 0) {
      writer.uint32(65).double(message.latitude);
    }
    if (message.longitude !== 0) {
      writer.uint32(73).double(message.longitude);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RegisterOrderRequestAddress {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRegisterOrderRequestAddress();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.address = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.neighborhood = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.number = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.complement = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.city = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.state = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.zipCode = reader.string();
          continue;
        }
        case 8: {
          if (tag !== 65) {
            break;
          }

          message.latitude = reader.double();
          continue;
        }
        case 9: {
          if (tag !== 73) {
            break;
          }

          message.longitude = reader.double();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseRegisterOrderRequest(): RegisterOrderRequest {
  return { userId: "", distance: 0, address: undefined };
}

export const RegisterOrderRequest: MessageFns<RegisterOrderRequest> = {
  encode(message: RegisterOrderRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.userId !== "") {
      writer.uint32(10).string(message.userId);
    }
    if (message.distance !== 0) {
      writer.uint32(17).double(message.distance);
    }
    if (message.comment !== undefined) {
      writer.uint32(26).string(message.comment);
    }
    if (message.paymentMethod !== undefined) {
      writer.uint32(34).string(message.paymentMethod);
    }
    if (message.address !== undefined) {
      RegisterOrderRequestAddress.encode(message.address, writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RegisterOrderRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRegisterOrderRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 17) {
            break;
          }

          message.distance = reader.double();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.comment = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.paymentMethod = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.address = RegisterOrderRequestAddress.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseOrderResponse(): OrderResponse {
  return { message: "", order: undefined };
}

export const OrderResponse: MessageFns<OrderResponse> = {
  encode(message: OrderResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.message !== "") {
      writer.uint32(10).string(message.message);
    }
    if (message.order !== undefined) {
      Order.encode(message.order, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): OrderResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseOrderResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.message = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.order = Order.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseAddress(): Address {
  return {
    id: "",
    orderId: "",
    address: "",
    neighborhood: "",
    number: "",
    complement: "",
    city: "",
    state: "",
    zipCode: "",
    latitude: 0,
    longitude: 0,
    createdAt: "",
    updatedAt: "",
  };
}

export const Address: MessageFns<Address> = {
  encode(message: Address, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    if (message.orderId !== "") {
      writer.uint32(18).string(message.orderId);
    }
    if (message.address !== "") {
      writer.uint32(26).string(message.address);
    }
    if (message.neighborhood !== "") {
      writer.uint32(34).string(message.neighborhood);
    }
    if (message.number !== "") {
      writer.uint32(42).string(message.number);
    }
    if (message.complement !== "") {
      writer.uint32(50).string(message.complement);
    }
    if (message.city !== "") {
      writer.uint32(58).string(message.city);
    }
    if (message.state !== "") {
      writer.uint32(66).string(message.state);
    }
    if (message.zipCode !== "") {
      writer.uint32(74).string(message.zipCode);
    }
    if (message.latitude !== 0) {
      writer.uint32(81).double(message.latitude);
    }
    if (message.longitude !== 0) {
      writer.uint32(89).double(message.longitude);
    }
    if (message.createdAt !== "") {
      writer.uint32(98).string(message.createdAt);
    }
    if (message.updatedAt !== "") {
      writer.uint32(106).string(message.updatedAt);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Address {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAddress();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.orderId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.address = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.neighborhood = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.number = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.complement = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.city = reader.string();
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.state = reader.string();
          continue;
        }
        case 9: {
          if (tag !== 74) {
            break;
          }

          message.zipCode = reader.string();
          continue;
        }
        case 10: {
          if (tag !== 81) {
            break;
          }

          message.latitude = reader.double();
          continue;
        }
        case 11: {
          if (tag !== 89) {
            break;
          }

          message.longitude = reader.double();
          continue;
        }
        case 12: {
          if (tag !== 98) {
            break;
          }

          message.createdAt = reader.string();
          continue;
        }
        case 13: {
          if (tag !== 106) {
            break;
          }

          message.updatedAt = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
}